Notes: Extract components and rename to Notes domain
